Add initial version of project files
This commit introduces the initial files, which include .gitignore, CSS and HTML files for the layout and styling of the webpage, README.md, and a style-guide file. The .gitignore file helps avoid accidental uploads of design files or unnecessary files. The CSS and HTML files establish a basic layout and design, while README.md provides an overview of the project.